requirements.txt
Flask==2.2.5
tensorflow==2.13.0  # This includes Keras, so no need for a separate keras package
numpy==1.23.5
Pillow==9.1.1
